refactor(test_support): move vector algebra methods to separate module
Split out MethodVectorAlgebraReport and method_vector_algebra into a new
file decomposition_vector_algebra.rs and update imports accordingly. This
cleans up the decomposition.rs test support code by isolating vector algebra
helper functionality, improving modularity and maintainability.
